Made Navigation capable of having different icons depending on which arrays are passed to it. Made AdminNavBar.tsx file that passes the necessary icons for admin. Replaced all references to Navigation with references to AdminNavBar
