+++ client/src/components/CRM/Accounts/SearchAccount/userSearchmain.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Accounts from './Accounts';
-import Navigation from '../../Navigation';
+import AdminNavBar from '../../AdminNavBar';
 /**
  * @returns {object} userSearchmain - has Navigation
  *  and Accounts to reroute to other components
@@ -9,7 +9,7 @@ const userSearchmain = () => {
   return (
     <div className='flex flex-row'>
       <Accounts/>
-      <Navigation/>
+      <AdminNavBar/>
     </div>
   );
 };
