+++ client/src/components/CRM/Contacts/ContactMain.tsx
@@ -1,5 +1,5 @@
 import React, {ReactElement} from 'react';
-import Navigation from '../Navigation';
+import AdminNavBar from '../AdminNavBar';
 import Contacts from './Contacts';
 
 /**
@@ -9,7 +9,7 @@ import Contacts from './Contacts';
 const ContactMain = (): ReactElement => {
   return (
     <div className='flex flex-row'>
-      <Navigation />
+      <AdminNavBar />
       <Contacts />
     </div>
   );
