+++ client/src/components/CRM/Reporting/DonationSummaryReport/DonationSummaryReport.tsx
@@ -2,7 +2,7 @@ import React, {useState} from 'react';
 
 import FilterComponent from './FilterComponent';
 import ReportComponent from './ReportComponent';
-import Navigation from '../../Navigation';
+import AdminNavBar from '../../AdminNavBar';
 
 import {Box, Grid} from '@mui/material';
 
@@ -87,7 +87,7 @@ const DonationSummaryReport = () => {
           </Grid>
         </Box>
       </Box>
-      <Navigation />
+      <AdminNavBar />
     </div>
   );
 };
