+++ client/src/components/CRM/Reporting/DonationSummaryReport/DonationSummaryReportMain.tsx
@@ -2,7 +2,7 @@ import React, {useState} from 'react';
 
 import FilterComponent from './FilterComponent';
 import ReportComponent from './ReportComponent';
-import Navigation from '../../Navigation';
+import AdminNavBar from '../../AdminNavBar';
 
 import {Box, Grid} from '@mui/material';
 
@@ -88,7 +88,7 @@ const DonationSummaryReportMain = () => {
           </Grid>
         </Box>
       </Box>
-      <Navigation />
+      <AdminNavBar />
     </div>
   );
 };
