+++ client/src/components/Navigation.tsx
@@ -3,16 +3,16 @@ import React, {useState, useRef, useEffect} from 'react';
 import logo from '../../Logo/2011_Logo_white.png';
 import '../../Logo/logo.css';
 import {useAuth0} from '@auth0/auth0-react';
-import AuthNav from '../Authentication/auth-nav';
+import AuthNav from './Authentication/auth-nav';
 import PropTypes from 'prop-types';
 import {useNavigate} from 'react-router-dom';
-import AdminNavDropdown from '../Ticketing/ticketingmanager/AdminNavDropdown';
+import AdminNavDropdown from './Ticketing/ticketingmanager/AdminNavDropdown';
 /**
  * CRM Navigation bar
  *
  * @returns {React.ReactElement}
  */
-const Navigation = (): React.ReactElement => {
+const Navigation = ({buttons}): React.ReactElement => {
   const [profile, setProfile] = useState(false);
   const [showMobileMenu, setShowMobileMenu] = useState(false);
 
@@ -77,34 +77,6 @@ const Navigation = (): React.ReactElement => {
     symbol: PropTypes.string,
   };
 
-  const navButtons = [
-    {
-      name: 'Dashboard',
-      url: '/admin',
-      symbol: 'M4 6a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2H6a2 2 0 01-2-2V6zM14 6a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2h-2a2 2 0 01-2-2V6zM4 16a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2H6a2 2 0 01-2-2v-2zM14 16a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2h-2a2 2 0 01-2-2v-2z',
-    },
-    {
-      name: 'Accounts',
-      url: '/admin/accounts',
-      symbol: 'M17 20h5v-2a3 3 0 00-5.356-1.857M17 20H7m10 0v-2c0-.656-.126-1.283-.356-1.857M7 20H2v-2a3 3 0 015.356-1.857M7 20v-2c0-.656.126-1.283.356-1.857m0 0a5.002 5.002 0 019.288 0M15 7a3 3 0 11-6 0 3 3 0 016 0zm6 3a2 2 0 11-4 0 2 2 0 014 0zM7 10a2 2 0 11-4 0 2 2 0 014 0z',
-    },
-    {
-      name: 'Contacts',
-      url: '/admin/contacts',
-      symbol: 'M10 6H5a2 2 0 00-2 2v9a2 2 0 002 2h14a2 2 0 002-2V8a2 2 0 00-2-2h-5m-4 0V5a2 2 0 114 0v1m-4 0a2 2 0 104 0m-5 8a2 2 0 100-4 2 2 0 000 4zm0 0c1.306 0 2.417.835 2.83 2M9 14a3.001 3.001 0 00-2.83 2M15 11h3m-3 4h2',
-    },
-    {
-      name: 'Reporting',
-      url: '/admin/reporting',
-      symbol: 'M9 17v-2m3 2v-4m3 4v-6m2 10H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z',
-    },
-    {
-      name: 'Task',
-      url: '/admin/tasks',
-      symbol: 'M12 6V4m0 2a2 2 0 100 4m0-4a2 2 0 110 4m-6 8a2 2 0 100-4m0 4a2 2 0 110-4m0 4v2m0-6V4m6 6v10m6-2a2 2 0 100-4m0 4a2 2 0 110-4m0 4v2m0-6V4',
-    },
-  ];
-
   const NavSymbols = ({navButtons}) => {
     return (
       <ul className="space-y-7 mt-30 ml-[rem] flex flex-col items-center mb-7">
@@ -140,7 +112,7 @@ const Navigation = (): React.ReactElement => {
               <img src={logo} className="mt-[5rem] w-[6rem]" alt="/" />
             </button>
           </div>
-          <NavSymbols navButtons={navButtons}/>
+          <NavSymbols navButtons={buttons}/>
         </div>
 
 
@@ -192,7 +164,7 @@ const Navigation = (): React.ReactElement => {
                     </svg>
                   </button>
                 </div>
-                  <NavSymbols navButtons={navButtons}/>
+                  <NavSymbols navButtons={buttons}/>
               </div>
 
               {/* Bottom corner nav elements */}
@@ -331,6 +303,10 @@ const Navigation = (): React.ReactElement => {
       </div>
     </div>
   );
+};
+
+Navigation.propTypes = {
+  buttons: PropTypes.array,
 };
 
 export default Navigation;
